pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "cbd"
version = "0.1.0"
description = "Baseline Carolina Bays detection MVP"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
  { name = "Project authors" }
]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "PyYAML>=6.0.2",
  "pydantic>=2.8.2",
  "geopandas>=1.0.1",
  "shapely>=2.0.5",
  "pyogrio>=0.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "ruff>=0.6.2",
  "mypy>=1.11.1",
]

[project.scripts]
cbd = "cbd.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/cbd"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_configs = true
ignore_missing_imports = true
